Remove unfinished and incomplete code.
