+++ src/_impl_extended_boolean.rs
@@ -1,8 +1,7 @@
-use crate::trap_spaces::ExtendedBoolean;
+use crate::ExtendedBoolean;
 use std::cmp::Ordering;
 use std::fmt::{Debug, Display, Formatter};
 use ExtendedBoolean::{Any, One, Zero};
-
 impl Debug for ExtendedBoolean {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         match self {
@@ -21,6 +20,7 @@ impl Display for ExtendedBoolean {
 
 /// Implements a "set-like" ordering for extended Booleans. "Any" is greater than constant values
 /// (since it covers both values), while constant values are incomparable between each other.
+
 impl PartialOrd for ExtendedBoolean {
     fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
         match (self, other) {
@@ -43,6 +43,7 @@ impl ExtendedBoolean {
 }
 
 /// Logical operations on extended booleans.
+
 impl ExtendedBoolean {
     pub fn negate(self) -> Self {
         match self {
