+++ src/_impl_space.rs
@@ -1,10 +1,8 @@
-use crate::trap_spaces::ExtendedBoolean::Any;
-use crate::trap_spaces::{ExtendedBoolean, OptionalExtendedBoolean, PartialSpace, Space};
-use crate::{BooleanNetwork, VariableId};
+use crate::ExtendedBoolean::{Any, One, Zero};
+use crate::{BooleanNetwork, ExtendedBoolean, Space, VariableId};
 use std::cmp::Ordering;
 use std::fmt::{Display, Formatter};
 use std::ops::{Index, IndexMut};
-use ExtendedBoolean::{One, Zero};
 
 impl Index<VariableId> for Space {
     type Output = ExtendedBoolean;
@@ -81,23 +79,7 @@ impl Space {
     }
 
     pub fn count_any(&self) -> usize {
-        self.0
-            .iter()
-            .filter(|it| **it == ExtendedBoolean::Any)
-            .count()
-    }
-
-    pub fn to_partial_space(self) -> PartialSpace {
-        PartialSpace(
-            self.0
-                .into_iter()
-                .map(|it| match it {
-                    One => OptionalExtendedBoolean::One,
-                    Zero => OptionalExtendedBoolean::Zero,
-                    ExtendedBoolean::Any => OptionalExtendedBoolean::Any,
-                })
-                .collect(),
-        )
+        self.0.iter().filter(|it| **it == Any).count()
     }
 
     /// Check if this `Space` is a **trap** space within the given `BooleanNetwork`.
@@ -128,36 +110,4 @@ impl Space {
 
         true
     }
-
-    /*
-        /// Check if this `Space` has a **trap conflict variable** (within the given `BooleanNetwork`).
-        ///
-        /// A "conflict variable" is a specific BN variable that can always be updated, and where the
-        /// update always leads outside of the given `Space`. In other words, from every state, one
-        /// can escape this `Space` in a single step using this variable.
-        ///
-        /// Note that a space with a conflict variable cannot be a trap space. However, a space that
-        /// is not a trap space may not have a single conflict variable, even if every state can
-        /// escape this `Space` in one step.
-        pub fn has_trap_conflict_variable(&self, network: &BooleanNetwork) -> Option<VariableId> {
-            for var in network.variables() {
-                let expected_value = self[var];
-                if expected_value.is_fixed() {
-                    if let Some(update) = network.get_update_function(var) {
-                        let actual_value = update.eval_in_space(self);
-                        if actual_value.is_fixed() && actual_value != expected_value {
-                            return Some(var);
-                        } else {
-                            // If actual value is "any", then the space is not a trap space, but
-                            // the variable is not transient either.
-                        }
-                    } else {
-                        // If the function is unknown, then this is also not a trap space, but
-                        // it does not guarantee that the variable is transient.
-                    }
-                }
-            }
-            None
-        }
-    */
 }
