+++ src/bin/bench_fixed_points_solver.rs
@@ -1,8 +1,8 @@
 use biodivine_lib_param_bn::biodivine_std::bitvector::BitVector;
 use biodivine_lib_param_bn::fixed_points::FixedPoints;
 use biodivine_lib_param_bn::solver_context::BnSolverContext;
-use biodivine_lib_param_bn::trap_spaces::Space;
 use biodivine_lib_param_bn::BooleanNetwork;
+use biodivine_lib_param_bn::Space;
 
 fn main() {
     let args = Vec::from_iter(std::env::args());
